+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/alarm/AlarmMessage.java
@@ -18,12 +18,12 @@
 
 package org.apache.skywalking.oap.server.core.alarm;
 
+import com.google.gson.JsonObject;
 import java.util.HashSet;
 import java.util.Set;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.skywalking.oap.server.core.analysis.manual.searchtag.Tag;
-
 import java.util.List;
 
 /**
@@ -43,4 +43,6 @@ public class AlarmMessage {
     private long startTime;
     private transient int period;
     private Set<String> hooks = new HashSet<>();
+    private String expression;
+    private JsonObject mqeMetricsSnapshot;
 }