+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/alarm/AlarmRecord.java
@@ -18,6 +18,7 @@
 
 package org.apache.skywalking.oap.server.core.alarm;
 
+import com.google.gson.Gson;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.skywalking.oap.server.core.analysis.Stream;
@@ -34,9 +35,7 @@ import org.apache.skywalking.oap.server.core.storage.annotation.SQLDatabase;
 import org.apache.skywalking.oap.server.core.storage.type.Convert2Entity;
 import org.apache.skywalking.oap.server.core.storage.type.Convert2Storage;
 import org.apache.skywalking.oap.server.core.storage.type.StorageBuilder;
-
 import java.util.List;
-
 import static org.apache.skywalking.oap.server.core.storage.StorageData.TIME_BUCKET;
 import static org.apache.skywalking.oap.server.core.source.DefaultScopeDefine.ALARM;
 
@@ -47,7 +46,7 @@ import static org.apache.skywalking.oap.server.core.source.DefaultScopeDefine.AL
 @SQLDatabase.ExtraColumn4AdditionalEntity(additionalTable = AlarmRecord.ADDITIONAL_TAG_TABLE, parentColumn = TIME_BUCKET)
 @BanyanDB.TimestampColumn(AlarmRecord.START_TIME)
 public class AlarmRecord extends Record {
-
+    private static final Gson GSON = new Gson();
     public static final String INDEX_NAME = "alarm_record";
     public static final String ADDITIONAL_TAG_TABLE = "alarm_record_tag";
     public static final String SCOPE = "scope";
@@ -59,6 +58,7 @@ public class AlarmRecord extends Record {
     public static final String RULE_NAME = "rule_name";
     public static final String TAGS = "tags";
     public static final String TAGS_RAW_DATA = "tags_raw_data";
+    public static final String SNAPSHOT = "snapshot";
 
     @Override
     public StorageID id() {
@@ -92,6 +92,8 @@ public class AlarmRecord extends Record {
     private List<String> tagsInString;
     @Column(name = TAGS_RAW_DATA, storageOnly = true, length = Tag.TAG_LENGTH)
     private byte[] tagsRawData;
+    @Column(name = SNAPSHOT, storageOnly = true, length = 50000)
+    private String snapshot;
 
     public static class Builder implements StorageBuilder<AlarmRecord> {
         @Override
@@ -106,6 +108,7 @@ public class AlarmRecord extends Record {
             record.setTimeBucket(((Number) converter.get(TIME_BUCKET)).longValue());
             record.setRuleName((String) converter.get(RULE_NAME));
             record.setTagsRawData(converter.getBytes(TAGS_RAW_DATA));
+            record.setSnapshot((String) converter.get(SNAPSHOT));
             // Don't read the TAGS as they are only for query.
             return record;
         }
@@ -122,6 +125,7 @@ public class AlarmRecord extends Record {
             converter.accept(RULE_NAME, storageData.getRuleName());
             converter.accept(TAGS_RAW_DATA, storageData.getTagsRawData());
             converter.accept(TAGS, storageData.getTagsInString());
+            converter.accept(SNAPSHOT, storageData.getSnapshot());
         }
     }
 }