+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/mqe/MQEValues.java
@@ -17,19 +17,15 @@
  * under the License.
  */
 
-package org.apache.skywalking.mqe.rt.type;
+package org.apache.skywalking.oap.server.core.query.mqe;
 
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
 import lombok.Data;
-import org.apache.skywalking.oap.server.core.query.type.KeyValue;
 
 @Data
-public class Metadata {
-    private List<KeyValue> labels  = new ArrayList<>();
+public class MQEValues {
+    private Metadata metric = new Metadata();
 
-    public void sortLabelsByKey(Comparator<String> comparator) {
-        labels.sort(Comparator.comparing(KeyValue::getKey, comparator));
-    }
+    private List<MQEValue> values = new ArrayList<>();
 }